Split test_intel_amx_attention_backend.py to pass CI of timeout
